requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
passlib[bcrypt]==1.7.4
crewai==0.1.0
streamlit==1.28.0
requests==2.31.0
pytest==7.4.0
pandas==2.0.3
fpdf==1.7.2
google-generativeai==0.3.0
schedule==1.2.0
tensorflow==2.16.1
matplotlib==3.7.2
scikit-learn==1.3.0
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
kaggle==1.5.16
langchain>=0.0.335
langchain-google-genai>=0.0.1